Draft wording for modifications
Reword the "multi-pass guarantee" so that it does not require rvalue iterators to be incrementable.

Sortable need not specify ForwardIterator<I> since Permutable<I> already does.

Declare is_sized_range and is_view normatively so the number of template parameters and class-key are fixed. Define view_base normatively so that implementors can't make it non-empty and/or virtual. Add stronger language for is_sized_range specialization.

Replace "semantically identical" in the definition of Boolean with explicit requirements