Removed casts to void
The casts are needed in Ranges v3, because the constrains are separated with commas. In C++ concepts, the constraints are separated by semicolons, so the cast to void adds nothing.
Also, the cast to void does not indicate that "we ignore the return value". It says "the result of swap should be castable to void", which is of no value.